Delete ReserveSlice and add ReserveCourse action to reservationsSlice, create citiesSlice, rearrange reducers and selectors, and refactor useEffect of reserve and reservations pages
