+++ src/app/store.js
@@ -3,16 +3,16 @@ import logger from 'redux-logger';
 import classesReducer from '../features/classes/classesSlice';
 import reservationsReducer from '../features/reservations/reservationsSlice';
 import sessionReducer from '../auth/sessionSlice';
-import reservePageReducer from '../features/reserve/reservePageSlice';
+import citiesReducer from '../features/reserve/CitiesSlice';
 import classDetailsReducer from '../features/ClassDetails/classDetailsSlice';
 
 const store = configureStore({
   reducer: {
     classes: classesReducer,
     classDetails: classDetailsReducer,
+    cities: citiesReducer,
     reservations: reservationsReducer,
     users: sessionReducer,
-    reserve: reservePageReducer,
   },
   middleware: (getDefaultMiddleware) => getDefaultMiddleware().concat(logger),
 });
