+++ src/features/reservations/reservationsSlice.js
@@ -9,12 +9,23 @@ const initialState = {
 };
 
 export const fetchReservations = createAsyncThunk(
-  'reservationsSlice/reservations', async (thunkAPI) => {
+  'reservationsSlice/getReservations', async (thunkAPI) => {
     try {
       const response = await expertClassApi.get('reservations');
       return response.data;
     } catch (error) {
       return thunkAPI.rejectWithValue({ error: error.message });
+    }
+  },
+);
+
+export const reserveCourse = createAsyncThunk(
+  'reservationsSlice/createReservation', async (formData, { rejectWithValue }) => {
+    try {
+      const response = await expertClassApi.post('reservations', { reservation: formData });
+      return response.data;
+    } catch (error) {
+      return rejectWithValue({ error: 'Unable to make a reservation. Please try again.' });
     }
   },
 );
@@ -29,16 +40,33 @@ export const reservationsSlice = createSlice({
         state.status = 'loading';
       })
       .addCase(fetchReservations.fulfilled, (state, action) => {
-        state.status = 'idle';
+        state.status = 'fulfilled';
         state.reservations = action.payload;
       })
       .addCase(fetchReservations.rejected, (state) => {
         state.status = 'rejected';
         state.error = 'Error fetching reservations';
+      })
+      .addCase(reserveCourse.pending, (state) => {
+        state.status = 'loading';
+      })
+      .addCase(reserveCourse.fulfilled, (state, action) => {
+        if (action.payload.status === 200) {
+          state.status = 'fulfilled';
+          state.reservations.push(action.payload.reservation);
+        } else {
+          state.status = 'fulfilled';
+          state.error = action.payload.message;
+        }
+      })
+      .addCase(reserveCourse.rejected, (state) => {
+        state.status = 'rejected';
+        state.error = 'Error making reservation';
       });
   },
 });
 
 export const userReservations = (state) => state.reservations.reservations;
+export const reservationsState = (state) => state.reservations;
 
 export default reservationsSlice.reducer;
